Cleaned up all errors, merged FilterDialog with Topbar for added functionaility albeit longer file. FIxed typing for everything, TODO: add delay and fade to popups, add loading screen in between changing genes, fix button to load gene
